Arrumado alguns problemas de visibilidade do spring security e implementado algumas páginas e mudanças feitas pela Lorene.

+++ ProjetoCruzeirao/src/main/java/sistema/managedbeans/CampeonatoMB.java
@@ -1,28 +1,75 @@
 package sistema.managedbeans;
 
-import java.util.*;
+import java.io.Serializable;
+import java.util.List;
 
+import javax.faces.application.FacesMessage;
 import javax.faces.bean.ManagedBean;
-import javax.faces.bean.SessionScoped;
+import javax.faces.bean.ViewScoped;
+import javax.faces.context.FacesContext;
+
+import org.primefaces.event.RowEditEvent;
 
 import sistema.entidades.Campeonato;
 import sistema.entidades.Categoria;
 import sistema.entidades.Juiz;
 import sistema.entidades.Local;
 import sistema.services.CampeonatoService;
+import sistema.services.UsuarioService;
 
-@SessionScoped
+@ViewScoped
 @ManagedBean
-public class CampeonatoMB {
+public class CampeonatoMB implements Serializable{
+	
+	/**
+	 * 
+	 */
+	private static final long serialVersionUID = 1L;
 	private Campeonato campeonato = new Campeonato();
 	private CampeonatoService campeonatoservice = new CampeonatoService();
 	private Local auxLocal;
 	private Juiz auxJuiz;
 	private Categoria auxCategoria;
+	private List<Campeonato> campeonatos;
+	
+	public void onRowEdit(RowEditEvent event) {
+
+		Campeonato c = ((Campeonato) event.getObject());
+		campeonatoservice.alterar(c);
+	}
 	
 	public void salvar() {
-		campeonato.setIdCampeonato(0);
-		campeonatoservice.salvar(campeonato);
+		
+		int count = 0;
+		
+		if (this.getCampeonatos() != null)
+			for (Campeonato c : this.getCampeonatos())
+				if (c.getNome().equals(campeonato.getNome()))
+					count++;
+		
+		if (campeonato.getDataInicioCampeonato().after(campeonato.getDataFimCampeonato())  || campeonato.getDataInicioInscricao().after(campeonato.getDataFimInscricao())) 
+		{
+			FacesContext context = FacesContext.getCurrentInstance();
+			context.addMessage(null, new FacesMessage(FacesMessage.SEVERITY_INFO, "Datas inválidas! A data fim do campeonato e do fim do prazo de inscrição não pode ocorrer antes das datas de início do campeonato e do prazo de inscrição, respectivamente", null));
+		}
+		
+		else if (count > 0) {
+			FacesContext context = FacesContext.getCurrentInstance();
+			context.addMessage(null, new FacesMessage(FacesMessage.SEVERITY_INFO, "Já existe um campeonato com esse nome!", null));
+		}
+		
+		else
+		{
+			campeonato = campeonatoservice.salvar(campeonato);
+			
+			if (campeonatos != null) {
+				campeonatos.add(campeonato);			
+				FacesContext context = FacesContext.getCurrentInstance();
+				context.addMessage(null, new FacesMessage(FacesMessage.SEVERITY_INFO, "Campeonato " + campeonato.getNome() + " cadastrado com sucesso!", null));
+			}
+
+		}	
+
 		campeonato = new Campeonato();
 	}
 	
@@ -31,12 +78,43 @@ public class CampeonatoMB {
 	}
 	
 	public void adicionarJuiz() {
-		campeonato.getJuizes().add(auxJuiz);
-		auxJuiz.getUsuario().getCampeonatos().add(campeonato);
+	
+		int count = 0;
+		
+		if (campeonato.getJuizes() != null)
+			for (Juiz j : campeonato.getJuizes())
+				if (j.getIdJuiz() == auxJuiz.getIdJuiz())
+					count++;
+		
+		if (count == 0) {
+			campeonato.getJuizes().add(auxJuiz);
+			auxJuiz.getUsuario().getCampeonatos().add(campeonato);		
+		}
+
+		
+		else {
+			FacesMessage mensagem = new FacesMessage("Juiz " + auxJuiz.getUsuario().getNome() + " já foi adicionado!");
+			FacesContext.getCurrentInstance().addMessage(null, mensagem);	
+		}
+		
 	}
 	
 	public void adicionarLocal() {
-		campeonato.getLocais().add(auxLocal);
+		
+		int count = 0;
+		
+		if (campeonato.getLocais() != null)
+			for (Local l : campeonato.getLocais())
+				if (l.getIdLocal() == auxLocal.getIdLocal())
+					count++;
+		
+		if (count == 0)
+			campeonato.getLocais().add(auxLocal);
+		
+		else {
+			FacesMessage mensagem = new FacesMessage("Local " + auxLocal.getNome() + " já foi adicionado!");
+			FacesContext.getCurrentInstance().addMessage(null, mensagem);	
+		}
 	}
 	
 	public void removerLocal(Local local) {
@@ -53,10 +131,25 @@ public class CampeonatoMB {
 	}
 	
 	public void remover(Campeonato campeonato) {
-		campeonatoservice.remover(campeonato);
-		for (Categoria aux : campeonato.getCategorias()) {
-			aux.setCampeonato(null);
+		
+		
+		if (campeonatoservice.pesquisarLocaisCampeonato(campeonato).size() > 0 || campeonatoservice.pesquisarJuizesCampeonato(campeonato).size() > 0 || campeonatoservice.pesquisarCategoriasCampeonato(campeonato).size() > 0) {
+			FacesContext context = FacesContext.getCurrentInstance();
+			context.addMessage(null, new FacesMessage(FacesMessage.SEVERITY_INFO, "Não é possível remover o campeonato, pois há local (is), ou juiz (es) ou categoria (as) amarrados a ele.", null));
 		}
+		
+		
+		else {
+			campeonatoservice.remover(campeonato);
+			campeonatos.remove(campeonato);
+		
+			for (Categoria aux : campeonato.getCategorias())
+				aux.setCampeonato(null);
+		
+			FacesMessage mensagem = new FacesMessage("Campeonato " + campeonato.getNome() + " removido com sucesso!");
+			FacesContext.getCurrentInstance().addMessage(null, mensagem);	
+		}
+				
 	}	
 
 	public Local getAuxLocal() {
@@ -91,7 +184,10 @@ public class CampeonatoMB {
 		this.campeonato = campeonato;
 	}
 	
-	public ArrayList<Campeonato> getCampeonatos() {
-		return campeonatoservice.getCampeonatos();
+	public List<Campeonato> getCampeonatos() {
+		if (campeonatos == null)
+			campeonatos = campeonatoservice.getCampeonatos();
+		
+		return campeonatos;
 	}
 }
